refactor(services): extract JobControlFlags from JobStore (#451 step 3/5)
Wave 6.4, third of the JobStore decomposition. Pull the cooperative-
cancel + pause request flags out of `JobStore` into their own module.
Each concern (cancel — H-0011 / Issue #152; pause — P-0099 v3-20c) is an
in-memory `set[str]` under its own lock, mirrored to an on-disk flag
file (`<job_dir>/CANCEL` / `<job_dir>/PAUSE`) that a fresh subprocess
JobStore reads at the cancel-aware callback boundary.

New module `services/_job_control_flags.py` (214 lines) — `JobControlFlags`:
- `__init__(metadata: JobMetadataStore)` — holds the two sets + locks;
  uses the metadata store only for path resolution (`jobs_dir` for the
  staging tempfile, `path_for` for the flag files)
- `request_cancel` / `is_cancel_requested` / `clear_cancel` /
  `_cancel_flag_path` — moved verbatim
- `request_pause` / `is_pause_requested` / `clear_pause` /
  `_pause_flag_path` — moved verbatim

`JobStore.__init__` constructs `self._flags = JobControlFlags(self._meta)`;
drops `_cancel_requested` / `_cancel_lock` / `_pause_requested` /
`_pause_lock`; the six public flag methods are one-line delegations;
`_cancel_flag_path` / `_pause_flag_path` removed (no external callers).
`os` / `contextlib` imports dropped from `jobs.py` (now only used in the
moved methods).

The in-memory set stays the source of truth for same-process callers;
the file is an IPC channel only — the cancel/pause invariant tests
(`test_inv_slot_release.py` INV-5 monotonic, `test_inv_pause_keeps_slot.py`,
`test_job_state_transitions.py`, the `test_reg_0152` child-cancel
visibility test) pass unchanged — they assert through the public
`request_cancel` / `is_cancel_requested` / ... surface.

Pure refactor, no public API or behaviour change — Change Gate not
required. `jobs.py` 759 → 622 lines. No test changes needed.

Verified: full non-slow backend suite (1533 passed, unchanged), ruff /
mypy clean.

Refs #451 (step 3/5). Next: extract JobLineage (parent-lock + child /
tree queries).
